+++ service-implementation/src/main/java/com/liaison/mailbox/service/queue/kafka/QueueServiceConstants.java
@@ -14,14 +14,13 @@ public interface QueueServiceConstants {
 
     String KAFKA_RELAY_PRODUCER_STREAM = "kafka.relay.producer.stream";
     String KAFKA_RELAY_CONSUMER_STREAM = "kafka.relay.consumer.stream";
+    String KAFKA_LEGACY_RELAY_PRODUCER_STREAM = "kafka.legacyrelay.producer.stream";
+    String KAFKA_LEGACY_RELAY_CONSUMER_STREAM = "kafka.legacyrelay.consumer.stream";
     String KAFKA_FILE_EVENT_READER_CONSUMER_STREAM= "kafka.fileeventreader.consumer.stream";
     
     String KAFKA_TOPIC_NAME_CREATE_DEFAULT = "kafka.topic.name.create.default";
     String KAFKA_TOPIC_NAME_DELETE_DEFAULT = "kafka.topic.name.delete.default";
     
-    String KAFKA_TOPIC_NAME_CREATE_LOWSECURE = "kafka.topic.name.create.lowsecure";
-    String KAFKA_TOPIC_NAME_DELETE_LOWSECURE = "kafka.topic.name.delete.lowsecure";
-
     String KAFKA_TOPIC_NAME_FS_EVENT_DEFAULT = "kafka.topic.name.fs.event.default";
     String KAFKA_TOPIC_NAME_FS_EVENT_LOWSECURE = "kafka.topic.name.fs.event.lowsecure";